// src/services/ibgeLocales-client/types.ts
export type getStatesResponse = {
  id: number;
  sigla: string;
  nome: string;
  regiao: {
    id: number;
    sigla: string;
    nome: string;
  };
};

export enum HttpStatusCode {
  ok = 200,
  created = 201,
  noContent = 204,
  badRequest = 400,
  unauthorized = 401,
  forbidden = 403,
  notFound = 404,
  serverError = 500,
}

export type HttpResponse<T = any> = {
  statusCode: HttpStatusCode;
  body?: T;
  isSucceeded: boolean;
};

export type getCitiesOfState = {
  id: number;
  nome: string;
  municipio: any;
};